LayoutTests/platform/mac/accessibility/html-with-aria-label.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html id="html" aria-label="Test Webpage">
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

Nothing to see.

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that aria-label will become the accessible name of the AXWebArea when used on the HTML tag.");

    if (window.accessibilityController) {

        // Test the values in a multi-select list box.
        var root = accessibilityController.rootElement.childAtIndex(0);
        shouldBe("root.role", "'AXRole: AXWebArea'");
        shouldBe("root.description", "'AXDescription: Test Webpage'");

        document.getElementById("html").setAttribute("aria-label", "New Title");
        shouldBe("root.description", "'AXDescription: New Title'");
    }

</script>

<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
